Fix padding mask, limit file size, +

+++ datavis.h
@@ -11,17 +11,16 @@
 #include <QBitmap>
 #include <QMouseEvent>
 
-
 class DataVis : public QWidget
 {
     Q_OBJECT
 public:
     explicit DataVis(QWidget *parent = 0,
                      char *buf = NULL,
-                     int size = 0,
+                     long size = 0,
                      QImage::Format img_format = QImage::Format_RGB32);
-    void loadData(char*, int, QImage::Format img_format = QImage::Format_RGB32);
-    int calcPadding(int size) { return ((w*depth) - (size % (w*depth))); }
+    void loadData(char*, long, QImage::Format img_format = QImage::Format_RGB32);
+    int calcPadding(long size) { return ((w*depth) - (size % (w*depth))); }
     QPoint getPoint(QMouseEvent* me) { return label->mapFromParent(me->pos()); }
     int getNBytes() { return n_bytes; }
     int getPixelDepth() { return depth; }
@@ -39,7 +38,7 @@ private:
     void scaleImage(float);
     void loadNext(int, int);
     char *data;
-    int n_bytes;
+    long n_bytes;
     int padding;
     int depth;
     QImage::Format format;
